pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "OpenThoughts-Agent"
version = "0.1.0"
description = "OT-Agent HPC launch and data tooling"
requires-python = ">=3.10"
dependencies = [
    # Core HPC + data infrastructure deps
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml",
    "omegaconf",
    "wandb",
    "bs4",
    "huggingface_hub>=0.20.0",
    "datasets>=2.0.0",
    "supabase>=2.22.3",
    "python-dotenv>=1.0.0",
    "harbor @ git+https://github.com/laude-institute/harbor.git@4aaebe4abfd005382870ce88b5b2670cd9afb54e",
    "dynamic-semaphore @ git+https://github.com/penfever/dynamic-semaphore",
    "google-cloud-storage",
    "h5py",
    "certifi",
    "rapidfuzz",
    "huggingface_hub",
    "uv>=0.4.17",
]

[project.optional-dependencies]
datagen-cpu = [
    "google-cloud-aiplatform>=1.67.0",
    "google-genai>=0.1.0",
]
datagen = [
    "ray[default]>=2.50.0",
    "torch-c-dlpack-ext==0.1.4",
    "google-cloud-aiplatform>=1.67.0",
    "google-genai>=0.1.0",
    "nvidia-cuda-nvcc-cu12==12.8.93; platform_system != 'Darwin'",
    "vllm==0.11.2; platform_system != 'Darwin'",
    "torch==2.9.0+cu128; platform_system != 'Darwin' or platform_machine != 'arm64'",
    "torch==2.9.0; platform_system == 'Darwin' and platform_machine == 'arm64'",
]
datagen-swesmith = [
    "swesmith",
]
cloud = [
    "skypilot>=0.6.1",
]
rl = [
    "skyrl @ git+https://github.com/mlfoundations/skyrl.git",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "data",
    "data.*",
    "database",
    "database.*",
    "sft",
    "sft.*",
    "eval",
    "eval.*",
    "hpc",
    "hpc.*",
    "rl",
    "rl.*",
    "scripts",
    "scripts.*",
    "train",
    "train.*",
]
exclude = [
    "docs*",
    "notes*",
    "wheels*",
    "tmp*",
]

[tool.uv.pip]
extra-index-url = ["https://download.pytorch.org/whl/cu128"]
[tool.uv]
index-strategy = "unsafe-best-match"
